Replaced hardcoded source actions by user defined ISourceAction.
