+++ mods/cnc/installer/covertops.yaml
@@ -3,5 +3,5 @@ covertops: Covert Operations Expansion (English)
 		GAME/GAME.DAT: be5a6c4c0a581da09e8f34a3bbf7bd17e525085c
 		CONQUER.MIX: 713b53fa4c188ca9619c6bbeadbfc86513704266
 	Install:
-		copy: .
+		Copy: .
 			^SupportDir|Content/cnc/scores-covertops.mix: SCORES.MIX
