+++ mods/cnc/installer/firstdecade.yaml
@@ -3,7 +3,7 @@ tfd: C&C The First Decade (English)
 		data1.hdr: bef3a08c3fc1b1caf28ca0dbb97c1f900005930e
 		data1.cab: 12ad6113a6890a1b4d5651a75378c963eaf513b9
 	Install:
-		extract-iscab: data1.hdr
+		ExtractIscab: data1.hdr
 			Volumes:
 				2: data2.cab
 				3: data3.cab
@@ -20,7 +20,7 @@ tfd: C&C The First Decade (English)
 				^SupportDir|Content/cnc/transit.mix: CnC\\TRANSIT.MIX
 				^SupportDir|Content/cnc/scores-covertops.mix: CnC\covert\SCORES.MIX
 				^SupportDir|Content/cnc/movies.mix: CnC\\MOVIES.MIX
-		extract-raw: ^SupportDir|Content/cnc/movies.mix
+		ExtractRaw: ^SupportDir|Content/cnc/movies.mix
 			^SupportDir|Content/cnc/movies/airstrk.vqa:
 				Offset: 1266
 				Length: 4444442
@@ -333,4 +333,4 @@ tfd: C&C The First Decade (English)
 			^SupportDir|Content/cnc/movies/banner.vqa:
 				Offset: 751962554
 				Length: 2229408
-		delete: ^SupportDir|Content/cnc/movies.mix
+		Delete: ^SupportDir|Content/cnc/movies.mix
