+++ mods/cnc/installer/gdi95.yaml
@@ -3,7 +3,7 @@ gdi95: C&C Gold (GDI Disc, English)
 		DISK.WAV: 8bef9a6687c0fe0afd57c6561df58fa6e64f145d
 		CONQUER.MIX: 833e02a09aae694659eb312d3838367f681d1b30
 	Install:
-		copy: .
+		Copy: .
 			^SupportDir|Content/cnc/conquer.mix: CONQUER.MIX
 			^SupportDir|Content/cnc/desert.mix: DESERT.MIX
 			^SupportDir|Content/cnc/general.mix: GENERAL.MIX
@@ -11,7 +11,7 @@ gdi95: C&C Gold (GDI Disc, English)
 			^SupportDir|Content/cnc/sounds.mix: SOUNDS.MIX
 			^SupportDir|Content/cnc/temperat.mix: TEMPERAT.MIX
 			^SupportDir|Content/cnc/winter.mix: WINTER.MIX
-		extract-blast: INSTALL/SETUP.Z
+		ExtractBlast: INSTALL/SETUP.Z
 			^SupportDir|Content/cnc/speech.mix:
 				Offset: 10203213
 				Length: 603293
@@ -21,7 +21,7 @@ gdi95: C&C Gold (GDI Disc, English)
 			^SupportDir|Content/cnc/transit.mix:
 				Offset: 11078017
 				Length: 3724462
-		extract-raw: MOVIES.MIX
+		ExtractRaw: MOVIES.MIX
 			^SupportDir|Content/cnc/movies/visor.vqa:
 				Offset: 786
 				Length: 4407162
