+++ mods/ra/installer/cnc95.yaml
@@ -2,5 +2,5 @@ cnc95: C&C Gold (GDI or Nod Disc, English)
 	IDFiles:
 		CONQUER.MIX: 833e02a09aae694659eb312d3838367f681d1b30
 	Install:
-		copy: .
+		Copy: .
 			^SupportDir|Content/ra/v2/cnc/desert.mix: DESERT.MIX
