+++ mods/ra/installer/origin.yaml
@@ -6,7 +6,7 @@ ra-origin: C&C The Ultimate Collection (Origin version, English)
 	IDFiles:
 		REDALERT.MIX: 0e58f4b54f44f6cd29fecf8cf379d33cf2d4caef
 	Install:
-		copy: .
+		Copy: .
 			^SupportDir|Content/ra/v2/expand/2nd_hand.aud: 2nd_hand.aud
 			^SupportDir|Content/ra/v2/expand/araziod.aud: araziod.aud
 			^SupportDir|Content/ra/v2/expand/await.aud: await.aud
@@ -27,7 +27,7 @@ ra-origin: C&C The Ultimate Collection (Origin version, English)
 			^SupportDir|Content/ra/v2/expand/expand2.mix: EXPAND2.MIX
 			^SupportDir|Content/ra/v2/expand/hires1.mix: HIRES1.MIX
 			^SupportDir|Content/ra/v2/expand/lores1.mix: LORES1.MIX
-		extract-raw: REDALERT.MIX
+		ExtractRaw: REDALERT.MIX
 			^SupportDir|Content/ra/v2/hires.mix:
 				Offset: 650612
 				Length: 5817417
@@ -40,7 +40,7 @@ ra-origin: C&C The Ultimate Collection (Origin version, English)
 			^SupportDir|Content/ra/v2/speech.mix:
 				Offset: 23042864
 				Length: 2003464
-		extract-raw: MAIN.MIX
+		ExtractRaw: MAIN.MIX
 			^SupportDir|Content/ra/v2/movies/aagun.vqa:
 				Offset: 668669829
 				Length: 3295512
@@ -464,5 +464,5 @@ cnc-origin: Command & Conquer (Origin version, English)
 	IDFiles:
 		CONQUER.MIX: 833e02a09aae694659eb312d3838367f681d1b30
 	Install:
-		copy: .
+		Copy: .
 			^SupportDir|Content/ra/v2/cnc/desert.mix: DESERT.MIX
