+++ mods/ra/installer/soviet95.yaml
@@ -4,7 +4,7 @@ soviet: Red Alert 95 (Soviet Disc, English)
 			Length: 4096
 		INSTALL/REDALERT.MIX: 0e58f4b54f44f6cd29fecf8cf379d33cf2d4caef
 	Install:
-		extract-raw: INSTALL/REDALERT.MIX
+		ExtractRaw: INSTALL/REDALERT.MIX
 			^SupportDir|Content/ra/v2/hires.mix:
 				Offset: 650612
 				Length: 5817417
@@ -17,7 +17,7 @@ soviet: Red Alert 95 (Soviet Disc, English)
 			^SupportDir|Content/ra/v2/speech.mix:
 				Offset: 23042864
 				Length: 2003464
-		extract-raw: MAIN.MIX
+		ExtractRaw: MAIN.MIX
 			^SupportDir|Content/ra/v2/conquer.mix:
 				Offset: 236
 				Length: 2177047
