+++ mods/ts/installer/tibsun.yaml
@@ -3,9 +3,9 @@ tibsun: Tiberian Sun (GDI or Nod Disc, English)
 		README.TXT: 45745c4a0c888317ec900208a426472779c42bf7
 		AUTOPLAY.WAV: 2dfce5d00f98b641849c29942b651f4e98d30e30
 	Install:
-		copy: .
+		Copy: .
 			^SupportDir|Content/ts/scores.mix: SCORES.MIX
-		extract-raw: INSTALL/TIBSUN.MIX
+		ExtractRaw: INSTALL/TIBSUN.MIX
 			^SupportDir|Content/ts/cache.mix:
 				Offset: 300
 				Length: 169176
